+++ index.html
@@ -1,16 +1,27 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <title>🎱 행운의 로또</title>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
+    <title>UKKO 로또 1등 당첨 기원</title>
   </head>
-
   <body>
-    <div id="app">
-      <h1>🎱 행운의 로또</h1>
+    <header>
+      <h1 class="title">🎱 행운의 로또</h1>
+    </header>
+    <div class="app-box">
+      <section id="purchasing-money">
+        <h2 class="title">🎱 내 번호 당첨 확인 🎱</h2>
+        <div class="message">구입할 금액을 입력해주세요.</div>
+        <form id="money-form">
+          <input id="money" type="text" placeholder="금액" />
+          <button id="purchasing-button" type="submit">구입</button>
+        </form>
+      </section>
     </div>
-    <script type="module" src="./src/js/index.js"></script>
+    <footer>
+      <p id="copyright">Copyright 2023. woowacourse</p>
+    </footer>
   </body>
 </html>
