+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "javascript-lotto",
-  "type": "module",
   "version": "1.0.0",
   "description": "우아한테크코스 웹 프론트엔드 레벨1 로또 미션",
   "license": "MIT",
